Alerts for product actions are done

+++ src/pages/Products.tsx
@@ -1,5 +1,7 @@
 import React, { FC, useContext } from 'react';
 import { RouteComponentProps, Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
+import { toast } from 'react-toastify';
 
 import PageContainer from '../components/UI/PageContainer';
 import Wrapper from '../components/UI/Wrapper';
@@ -8,6 +10,7 @@ import ProductContext from '../context/Product/ProductContext';
 import Button from '../components/UI/Button';
 
 import { Product } from '../types/Api';
+import { ToastAlertOptions as options } from '../data/ToastAlert';
 
 interface ProductsProps extends RouteComponentProps{}
 
@@ -16,7 +19,24 @@ const Products: FC<ProductsProps> = props => {
     const { products, setEditingProduct, deleteProduct } = useContext(ProductContext);
 
     const deleteProductRequest = async (id: number) => {
-        await deleteProduct(id);
+        const promptResponse = await Swal.fire({
+            title: 'Eliminar proveedor',
+            icon: 'question',
+            text: '¿Estás seguro de que quieres eliminar este producto?',
+            confirmButtonText: 'Si, borrar',
+            confirmButtonColor: 'green',
+            showCancelButton: true,
+            cancelButtonText: 'No, cancelar',
+            cancelButtonColor: 'red',
+        });
+
+        if (promptResponse.value === promptResponse.isConfirmed) {
+            const response = await deleteProduct(id);
+            if (response.error)
+                toast.error(response.message, options);
+            else
+                toast.success(response.message, options);
+        }
     }
 
     const editProduct = (product: Product) => {
